Create PS-OTX-API-To-CSV.ps1
This file is the building block of a new public sector project used to keep a trusted client userbase updated and informed while utilizing and building upon the existing infrastructure available to date. We are attempting to group a majority of viable and reliable security related resources together in a manner such that a user may add it to their websites and pull down all pulses and IOC datasets available through other security-oriented agencies. We will be adding to this repo in order to complete the project and will create changelogs with each integration so as to keep our userbase informed. This repo has been left public to allow users the opportunity to contribute and or utilize the project to their security advantage.